Refactor to allow for better testing Move objects used in tests to test classes
