+++ src/Maestro/Microsoft.DotNet.Maestro.Tasks/src/PushMetadataToBuildAssetRegistry.cs
@@ -37,7 +37,7 @@ namespace Microsoft.DotNet.Maestro.Tasks
         private bool IsStableBuild { get; set; } = false;
 
         private bool IsReleaseOnlyPackageVersion { get; set; } = false;
-        
+
         public string RepoRoot { get; set; }
 
         public string AssetVersion { get; set; }
@@ -286,92 +286,105 @@ namespace Microsoft.DotNet.Maestro.Tasks
             string manifestsFolderPath,
             CancellationToken cancellationToken)
         {
-            var buildsManifestMetadata = new List<BuildData>();
-            var signingInfo = new List<SigningInformation>();
-            ManifestBuildData manifestBuildData = null;
+            List<Manifest> parsedManifests = new List<Manifest>();
 
             foreach (string manifestPath in Directory.GetFiles(manifestsFolderPath, SearchPattern, SearchOption.AllDirectories))
             {
                 cancellationToken.ThrowIfCancellationRequested();
 
-                var xmlSerializer = new XmlSerializer(typeof(Manifest));
-                using (var stream = new FileStream(manifestPath, FileMode.Open))
+                XmlSerializer xmlSerializer = new XmlSerializer(typeof(Manifest));
+                using FileStream stream = new FileStream(manifestPath, FileMode.Open);
+                Manifest manifest = (Manifest)xmlSerializer.Deserialize(stream);
+                parsedManifests.Add(manifest);
+            }
+
+            return ParseBuildManifestsMetadata(parsedManifests, cancellationToken);
+        }
+
+        internal (List<BuildData> buildData, List<SigningInformation> signingInformation, ManifestBuildData manifestBuildData) ParseBuildManifestsMetadata(
+            List<Manifest> parsedManifests,
+            CancellationToken cancellationToken)
+        {
+            var buildsManifestMetadata = new List<BuildData>();
+            var signingInfo = new List<SigningInformation>();
+            ManifestBuildData manifestBuildData = null;
+
+            foreach (Manifest manifest in parsedManifests)
+            {
+                cancellationToken.ThrowIfCancellationRequested();
+
+                if (manifestBuildData == null)
                 {
-                    var manifest = (Manifest)xmlSerializer.Deserialize(stream);
-
-                    if (manifestBuildData == null)
+                    manifestBuildData = new ManifestBuildData(manifest);
+                }
+                else
+                {
+                    if (!manifestBuildData.Equals(new ManifestBuildData(manifest)))
                     {
-                        manifestBuildData = new ManifestBuildData(manifest);
+                        throw new Exception("Attributes should be the same in all manifests.");
                     }
-                    else
+                }
+
+                var assets = new List<AssetData>();
+
+                foreach (Package package in manifest.Packages)
+                {
+                    AddAsset(
+                        assets,
+                        package.Id,
+                        package.Version,
+                        manifest.InitialAssetsLocation ?? manifest.Location,
+                        (manifest.InitialAssetsLocation == null) ? LocationType.NugetFeed : LocationType.Container,
+                        package.NonShipping);
+                }
+
+                foreach (Blob blob in manifest.Blobs)
+                {
+                    string version = GetVersion(blob.Id);
+
+                    if (string.IsNullOrEmpty(version))
                     {
-                        if (!manifestBuildData.Equals(new ManifestBuildData(manifest)))
-                        {
-                            throw new Exception("Attributes should be the same in all manifests.");
-                        }
-                    }
-
-                    var assets = new List<AssetData>();
-
-                    foreach (Package package in manifest.Packages)
-                    {
-                        AddAsset(
-                            assets,
-                            package.Id,
-                            package.Version,
-                            manifest.InitialAssetsLocation ?? manifest.Location,
-                            (manifest.InitialAssetsLocation == null) ? LocationType.NugetFeed : LocationType.Container,
-                            package.NonShipping);
-                    }
-
-                    foreach (Blob blob in manifest.Blobs)
-                    {
-                        string version = GetVersion(blob.Id);
-
-                        if (string.IsNullOrEmpty(version))
-                        {
-                            Log.LogWarning($"Version could not be extracted from '{blob.Id}'");
-                            version = string.Empty;
-                        }
-
-                        AddAsset(
-                            assets,
-                            blob.Id,
-                            version,
-                            manifest.InitialAssetsLocation ?? manifest.Location,
-                            LocationType.Container,
-                            blob.NonShipping);
-
-                        blobSet.Add(blob.Id);
+                        Log.LogWarning($"Version could not be extracted from '{blob.Id}'");
+                        version = string.Empty;
                     }
 
-                    // For now we aren't persisting this property, so we just record this info in the task scope
-                    IsStableBuild = bool.Parse(manifest.IsStable.ToLower());
+                    AddAsset(
+                        assets,
+                        blob.Id,
+                        version,
+                        manifest.InitialAssetsLocation ?? manifest.Location,
+                        LocationType.Container,
+                        blob.NonShipping);
 
-                    // The AzureDevOps properties can be null in the Manifest, but maestro needs them. Read them from the environment if they are null in the manifest.
-                    var buildInfo = new BuildData(
-                        commit: manifest.Commit,
-                        azureDevOpsAccount: manifest.AzureDevOpsAccount ?? GetAzDevAccount(),
-                        azureDevOpsProject: manifest.AzureDevOpsProject ?? GetAzDevProject(),
-                        azureDevOpsBuildNumber: manifest.AzureDevOpsBuildNumber ?? GetAzDevBuildNumber(),
-                        azureDevOpsRepository: manifest.AzureDevOpsRepository ?? GetAzDevRepository(),
-                        azureDevOpsBranch: manifest.AzureDevOpsBranch ?? GetAzDevBranch(),
-                        stable: IsStableBuild,
-                        released: false)
-                    {
-                        Assets = assets.ToImmutableList(),
-                        AzureDevOpsBuildId = manifest.AzureDevOpsBuildId ?? GetAzDevBuildId(),
-                        AzureDevOpsBuildDefinitionId = manifest.AzureDevOpsBuildDefinitionId ?? GetAzDevBuildDefinitionId(),
-                        GitHubRepository = manifest.Name,
-                        GitHubBranch = manifest.Branch,
-                    };
+                    blobSet.Add(blob.Id);
+                }
 
-                    buildsManifestMetadata.Add(buildInfo);
+                // For now we aren't persisting this property, so we just record this info in the task scope
+                IsStableBuild = bool.Parse(manifest.IsStable.ToLower());
 
-                    if (manifest.SigningInformation != null)
-                    {
-                        signingInfo.Add(manifest.SigningInformation);
-                    }
+                // The AzureDevOps properties can be null in the Manifest, but maestro needs them. Read them from the environment if they are null in the manifest.
+                var buildInfo = new BuildData(
+                    commit: manifest.Commit,
+                    azureDevOpsAccount: manifest.AzureDevOpsAccount ?? GetAzDevAccount(),
+                    azureDevOpsProject: manifest.AzureDevOpsProject ?? GetAzDevProject(),
+                    azureDevOpsBuildNumber: manifest.AzureDevOpsBuildNumber ?? GetAzDevBuildNumber(),
+                    azureDevOpsRepository: manifest.AzureDevOpsRepository ?? GetAzDevRepository(),
+                    azureDevOpsBranch: manifest.AzureDevOpsBranch ?? GetAzDevBranch(),
+                    stable: IsStableBuild,
+                    released: false)
+                {
+                    Assets = assets.ToImmutableList(),
+                    AzureDevOpsBuildId = manifest.AzureDevOpsBuildId ?? GetAzDevBuildId(),
+                    AzureDevOpsBuildDefinitionId = manifest.AzureDevOpsBuildDefinitionId ?? GetAzDevBuildDefinitionId(),
+                    GitHubRepository = manifest.Name,
+                    GitHubBranch = manifest.Branch,
+                };
+
+                buildsManifestMetadata.Add(buildInfo);
+
+                if (manifest.SigningInformation != null)
+                {
+                    signingInfo.Add(manifest.SigningInformation);
                 }
             }
 
@@ -658,7 +671,7 @@ namespace Microsoft.DotNet.Maestro.Tasks
                             Commit = GetAzDevCommit(),
                             IsStable = IsStableBuild.ToString(),
                             PublishingVersion = (PublishingInfraVersion)manifestBuildData.PublishingVersion,
-                            IsReleaseOnlyPackageVersion =manifestBuildData.IsReleaseOnlyPackageVersion
+                            IsReleaseOnlyPackageVersion = manifestBuildData.IsReleaseOnlyPackageVersion
 
                         });
